Reorganize tools
